Move existing stateIsActive tests to their own file
